+++ src/baobab_scryfall_api_caller/services/cards/__init__.py
@@ -1,0 +1,5 @@
+"""Exports publics du domaine Cards."""
+
+from baobab_scryfall_api_caller.services.cards.cards_service import CardsService
+
+__all__ = ["CardsService"]